Use floating-point values for pulse burst length

+++ libremote/src/main/kotlin/libremote/ir/IrTransmitter.kt
@@ -28,13 +28,13 @@ typealias IrTransmitFunction = (carrierFrequency: Int, pattern: Iterable<Int>) -
  */
 class IrTransmitter(
     private val carrierFrequency: Int,
-    private val pulseBurstLength: Int,
+    private val pulseBurstLength: Double,
     private val transmit: IrTransmitFunction
 ) : Transmitter {
     /**
      * @see [Transmitter.transmit]
      */
     override fun transmit(data: Iterable<Int>) {
-        transmit(carrierFrequency, data.map { it * pulseBurstLength })
+        transmit(carrierFrequency, data.map { (it * pulseBurstLength).toInt() })
     }
 }
